PRIdPTR failed in 64-bit mode.
git-svn-id: svn+ssh://svn.corp.yahoo.com/yahoo/yst/websearch/projects/clusterlib/trunk@255693 73f437aa-624d-0410-a3ad-9285e900ec26

+++ src/activenode/processslotupdater.cc
@@ -183,8 +183,8 @@ ProcessSlotUpdater::run()
             }
             else if (waitedPid == 0) {
                 LOG_DEBUG(CL_LOG, 
-                          "run: No update on process %" PRIdPTR, 
-                          oldPid);
+                          "run: No update on process %" PRId32, 
+                          static_cast<int32_t>(oldPid));
             }
             else if (waitedPid != oldPid) {
                 ostringstream oss;
